{cmake} Clean up of GL plugins
- add shader files to targets so they can be edited in an IDE as part of the project
- keep the shader path as a target property to be used when installing
- installing not clean up yet...
- move SSAO's RandomKit into an extern directory since it isn't our code (+4 squashed commits)
