+++ plugins/core/GL/qEDL/CMakeLists.txt
@@ -1,15 +1,17 @@
-option( PLUGIN_GL_QEDL "Check to install qEDL plugin" OFF )
+option( PLUGIN_GL_QEDL "Install qEDL plugin" OFF )
 
-# CloudCompare 'EDL' plugin
-if (PLUGIN_GL_QEDL)
+if( PLUGIN_GL_QEDL )
     project( QEDL_GL_PLUGIN )
     
-    AddPlugin( NAME ${PROJECT_NAME} TYPE gl )
-
-	include_directories( src )
+    AddPlugin(
+        NAME ${PROJECT_NAME}
+        TYPE gl
+        SHADER_FOLDER EDL
+    )
 	
+	add_subdirectory( include )
 	add_subdirectory( src )
+	add_subdirectory( shaders )
 	
-    set( CC_SHADER_FOLDER EDL )
     include( CMakePluginTemplate )
 endif()
