+++ plugins/core/GL/qEDL/include/CMakeLists.txt
@@ -1,8 +1,8 @@
 
 target_sources( ${PROJECT_NAME}
     PRIVATE
-        ${CMAKE_CURRENT_LIST_DIR}/Bilateral.h
-        ${CMAKE_CURRENT_LIST_DIR}/ExampleGLPlugin.h
+        ${CMAKE_CURRENT_LIST_DIR}/ccEDLFilter.h
+        ${CMAKE_CURRENT_LIST_DIR}/qEDL.h
 )
 
 target_include_directories( ${PROJECT_NAME}
