+++ plugins/core/GL/qEDL/include/ccEDLFilter.h
