+++ plugins/core/GL/qEDL/qEDL.qrc
@@ -1,6 +1,6 @@
 <RCC>
-  <qresource prefix="/CC/plugin/qEDL" >
-    <file>images/cc_edl.png</file>
-    <file>info.json</file>
-  </qresource>
+    <qresource prefix="/CC/plugin/qEDL">
+        <file>images/cc_edl.png</file>
+        <file>info.json</file>
+    </qresource>
 </RCC>
