+++ plugins/core/GL/qEDL/shaders/CMakeLists.txt
@@ -1,0 +1,7 @@
+target_sources( ${PROJECT_NAME}
+    PRIVATE
+        ${CMAKE_CURRENT_LIST_DIR}/EDL/edl_mix.frag
+        ${CMAKE_CURRENT_LIST_DIR}/EDL/edl_mix.vert
+        ${CMAKE_CURRENT_LIST_DIR}/EDL/edl_shade.frag
+        ${CMAKE_CURRENT_LIST_DIR}/EDL/edl_shade.vert
+)
