+++ plugins/core/GL/qEDL/src/CMakeLists.txt
@@ -1,12 +1,5 @@
-
-set( CC_PLUGIN_CUSTOM_HEADER_LIST
-	${CC_PLUGIN_CUSTOM_HEADER_LIST} 
-	${CMAKE_CURRENT_SOURCE_DIR}/ccEDLFilter.h
-	PARENT_SCOPE
-)
-
-set( CC_PLUGIN_CUSTOM_SOURCE_LIST
-	${CC_PLUGIN_CUSTOM_SOURCE_LIST} 
-	${CMAKE_CURRENT_SOURCE_DIR}/ccEDLFilter.cpp
-	PARENT_SCOPE
+target_sources( ${PROJECT_NAME}
+    PRIVATE
+        ${CMAKE_CURRENT_LIST_DIR}/ccEDLFilter.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/qEDL.cpp
 )
