+++ plugins/core/GL/qSSAO/extern/Randomkit/Randomkit/README
