+++ plugins/core/GL/qSSAO/extern/Randomkit/Randomkit/getopt.c
