+++ plugins/core/GL/qSSAO/extern/Randomkit/Randomkit/randomkit_example.c
