+++ plugins/core/GL/qSSAO/include/CMakeLists.txt
@@ -1,8 +1,8 @@
 
 target_sources( ${PROJECT_NAME}
     PRIVATE
-        ${CMAKE_CURRENT_LIST_DIR}/Bilateral.h
-        ${CMAKE_CURRENT_LIST_DIR}/ExampleGLPlugin.h
+       ${CMAKE_CURRENT_LIST_DIR}/ccSSAOFilter.h
+       ${CMAKE_CURRENT_LIST_DIR}/qSSAO.h
 )
 
 target_include_directories( ${PROJECT_NAME}
