+++ plugins/core/GL/qSSAO/include/ccSSAOFilter.h
