+++ plugins/core/GL/qSSAO/include/qSSAO.h
@@ -25,11 +25,11 @@ class qSSAO : public QObject, public ccGLPluginInterface
 {
 	Q_OBJECT
 	Q_INTERFACES(ccGLPluginInterface)
-	Q_PLUGIN_METADATA(IID "cccorp.cloudcompare.plugin.qSSAO" FILE "info.json")
+	Q_PLUGIN_METADATA(IID "cccorp.cloudcompare.plugin.qSSAO" FILE "../info.json")
 
 public:
 	explicit qSSAO(QObject* parent = nullptr);
-	
+
 	~qSSAO() override = default;
 
 	//inherited from ccGLFilterPluginInterface
