+++ plugins/core/GL/qSSAO/shaders/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources( ${PROJECT_NAME}
+    PRIVATE
+        ${CMAKE_CURRENT_LIST_DIR}/SSAO/ssao.frag
+        ${CMAKE_CURRENT_LIST_DIR}/SSAO/ssao.vert
+)
