+++ plugins/core/GL/qSSAO/src/ccSSAOFilter.cpp
