+++ plugins/example/ExampleGLPlugin/CMakeLists.txt
@@ -1,12 +1,16 @@
 option( PLUGIN_EXAMPLE_GL "Install example GL plugin" OFF )
 
-if ( PLUGIN_EXAMPLE_GL )
+if( PLUGIN_EXAMPLE_GL )
     project( ExampleGLPlugin )
     
-    AddPlugin( NAME ${PROJECT_NAME} TYPE gl )
-    
-    include( CMakePluginTemplate )
-    
+    AddPlugin(
+        NAME ${PROJECT_NAME}
+        TYPE gl
+    )
+        
     add_subdirectory( include )
     add_subdirectory( src )
+    add_subdirectory( shaders )
+    
+    include( CMakePluginTemplate )
 endif()
