+++ plugins/example/ExampleGLPlugin/shaders/CMakeLists.txt
@@ -1,0 +1,8 @@
+# Add shaders in a folder in this directory, then list the files here.
+# This will add them to the project so you can edit them in your IDE.
+
+#target_sources( ${PROJECT_NAME}
+#    PRIVATE
+#        ${CMAKE_CURRENT_LIST_DIR}/<folder>/<shader>.frag
+#        ${CMAKE_CURRENT_LIST_DIR}/<folder>/<shader>.vert
+#)
