+++ plugins/example/ExampleGLPlugin/src/CMakeLists.txt
@@ -2,4 +2,5 @@
 target_sources( ${PROJECT_NAME}
     PRIVATE
         ${CMAKE_CURRENT_LIST_DIR}/Bilateral.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/ExampleGLPlugin.cpp
 )
