+++ plugins/example/ExampleGLPlugin/src/ExampleGLPlugin.cpp
